ingredient

+++ lesson_02/personal/ingredient.py
@@ -2,7 +2,7 @@ total_no_of_cookies = 48
 cups_of_sugar = 1.5
 cups_of_butter = 1
 cups_of_flour = 2.75
-no_of_cookies_req = round(float(input("cookies required :")))
+no_of_cookies_req = round(int(input("cookies required :")))
 print("no of cookies required in roundup : {0:2f} ".format(no_of_cookies_req))
 req_sugar = (no_of_cookies_req / total_no_of_cookies) *  cups_of_sugar
 print("no of cups of sugar_required  : {0:.2f}".format(req_sugar))
@@ -10,6 +10,7 @@ req_butter = (no_of_cookies_req / total_no_of_cookies) * cups_of_butter
 print("no of cups of butter_required : {0:.2f}".format(req_butter))
 flour_req = (no_of_cookies_req / total_no_of_cookies) * cups_of_flour
 print("no of cups of flour_required : {0:.2f}".format(flour_req))
+
 
 #Question :
 #A cookie recipe calls for the following ingredients:
